+++ src/PagesLogin/Register.jsx
@@ -2,6 +2,7 @@ import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import '../Stylesheet/style.css'; 
+import NavigationBar from '../Navigator/Navigator';
 
 export default function Register() {
     
@@ -10,14 +11,10 @@ export default function Register() {
             <header>
                 <h1 className="titulo">Transportes la libertad<span>Agency</span></h1>
             </header>
-            <nav className="nav-principal" style={{ backgroundColor: '#b9fdfd' }}>
-            <a className="enlace" href="">Inicio</a>
-                <a className="enlace" href="">Administrador</a>
-                <a className="enlace" href="">Mecanico</a>
-                <a className="enlace" href="">Almacen</a>
-                <a className="enlace" href="">Registrar</a>
-                <a className="enlace" href="">Salir</a>
-            </nav>
+            <NavigationBar>
+
+            </NavigationBar>
+
             <div className="container">
                 <div className="col-lg-12">
                     <h1>Registrar</h1>
