+++ src/PagesStore/AddRequestStore.jsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import '../Stylesheet/style.css';
-import NavigationBar from './Navigator';
-export default function AddRequestStore(){
+import NavigationBar from '../Navigator/Navigator';
+
+export default function AddRequestStore() {
     return (
         <div>
             <header>
                 <h1 className="titulo">Transportes la libertad<span>Agency</span></h1>
             </header>
-        <NavigationBar>
+            <NavigationBar>
 
-        </NavigationBar>
-           
+            </NavigationBar>
 
             <div className="container">
                 <div className="col-lg-12">
@@ -19,15 +19,15 @@ export default function AddRequestStore(){
                     <form>
                         <div className="mb-3">
                             <label htmlFor="fecha" className="form-label">Fecha:</label>
-                            <input id="fecha" className="form-control" type="date" name="fecha"/>
+                            <input id="fecha" className="form-control" type="date" name="fecha" />
                         </div>
                         <div className="mb-3">
                             <label htmlFor="cantidad" className="form-label">Cantidad:</label>
-                            <input id="cantidad" className="form-control" type="text" name="cantidad"/>
+                            <input id="cantidad" className="form-control" type="text" name="cantidad" />
                         </div>
                         <div className="mb-3">
                             <label htmlFor="repuesto" className="form-label">Repuesto:</label>
-                            <input id="repuesto" className="form-control" type="text" name="repuesto"/>
+                            <input id="repuesto" className="form-control" type="text" name="repuesto" />
                         </div>
                         <button className="btn btn-success" type="submit">Agregar Solicitud</button>
                     </form>
@@ -37,6 +37,6 @@ export default function AddRequestStore(){
             <footer className="footer">
                 <p>Todos los derechos reservados 2024</p>
             </footer>
-       </div>
-);
+        </div>
+    );
 }
