--- src/pageAddReparationRequest/AddReparationRequest.css
@@ -1,81 +1,0 @@
-
-body {
-    font-family: Arial, sans-serif;
-  }
-  
-  .container {
-    width: 60%;
-    margin: 0 auto;
-    padding: 20px;
-  }
-  
-  header {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  h1 {
-    margin: 10px 0;
-  }
-  
-  nav {
-    display: flex;
-    justify-content: center;
-    gap: 10px;
-  }
-  
-  nav button {
-    padding: 10px 15px;
-    border: 1px solid #ccc;
-    background-color: #f4f4f4;
-    cursor: pointer;
-  }
-  
-  nav button.active {
-    background-color: #ddd;
-  }
-  
-  .form-section {
-    margin-top: 20px;
-  }
-  
-  h2 {
-    text-align: center;
-  }
-  
-  form {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    max-width: 300px;
-    margin: 0 auto;
-  }
-  
-  .input-group {
-    display: flex;
-    justify-content: space-between;
-  }
-  
-  .input-group label {
-    margin-right: 10px;
-  }
-  
-  input {
-    width: 100%;
-    padding: 5px;
-  }
-  
-  .submit-btn {
-    background-color: #007bff;
-    color: white;
-    padding: 10px;
-    border: none;
-    cursor: pointer;
-    width: 100%;
-  }
-  
-  .submit-btn:hover {
-    background-color: #0056b3;
-  }
-
