--- src/pageReparationRequest/ReparationRequest.css
@@ -1,62 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-  }
-  
-  .container {
-    text-align: center;
-    margin: 20px;
-  }
-  
-  .navbar {
-    margin-bottom: 20px;
-  }
-  
-  .nav-button {
-    margin: 5px;
-    padding: 10px 20px;
-    background-color: #f0f0f0;
-    border: 1px solid #ccc;
-    cursor: pointer;
-  }
-  
-  .nav-button:hover {
-    background-color: #ddd;
-  }
-  
-  .form-container {
-    background-color: #f9f9f9;
-    padding: 20px;
-    border-radius: 10px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    display: inline-block;
-  }
-  
-  .form-group {
-    margin-bottom: 15px;
-    text-align: left;
-  }
-  
-  label {
-    display: block;
-    margin-bottom: 5px;
-  }
-  
-  input {
-    padding: 10px;
-    width: 200px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-  }
-  
-  .submit-button {
-    padding: 10px 20px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-  }
-  
-  .submit-button:hover {
-    background-color: #0056b3;
-  }
