--- src/pageReportTable/ReporTable.css
@@ -1,84 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-  }
-  
-  .container {
-    text-align: center;
-    margin: 20px;
-  }
-  
-  .navbar {
-    margin-bottom: 20px;
-  }
-  
-  .nav-button {
-    margin: 5px;
-    padding: 10px 20px;
-    background-color: #f0f0f0;
-    border: 1px solid #ccc;
-    cursor: pointer;
-  }
-  
-  .nav-button:hover {
-    background-color: orange;
-  }
-  
-  .search-container {
-    margin-bottom: 20px;
-  }
-  
-  input {
-    padding: 10px;
-    width: 200px;
-    margin-right: 10px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-  }
-  
-  .search-button {
-    padding: 10px 20px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-  }
-  
-  .pdf-button {
-    padding: 10px 20px;
-    background-color: #dc3545;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    margin-left: 10px;
-    cursor: pointer;
-  }
-  
-  table {
-    width: 100%;
-    border-collapse: collapse;
-    margin: 20px auto;
-  }
-  
-  th, td {
-    padding: 10px;
-    text-align: left;
-    border: 1px solid #ddd;
-  }
-  
-  th {
-    background-color: #f4f4f4;
-  }
-  
-  .solicitar-button {
-    padding: 5px 10px;
-    background-color: #28a745;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-  }
-  
-  .solicitar-button:hover {
-    background-color: #218838;
-  }
